Addressed comments about Empty List/Optional
After making these changes, one of the tests failed. So in the case of
exceptions in refresh, we should still mark the refresh as true, so that
the hub will start. This will keep the symmetry that both the services
that have been running and the new instances would work similarly.

Signed-off-by: Pranith Kumar K <pranith.karampuri@phonepe.com>